license-server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
jinja2
bcrypt
